fix sign error
Fix the problem that the following devices cannot be configured due to an error when sending MQTT messages.

- Device : mss110 us mt7682 (hardware:2.0.0 firmware:2.1.19) 
- Error : "payload":{"error":{"code":5001,"detail":"sign error"}}

If this does not set the sign parameter correctly when sending an MQTT message, the device will not accept the request. The calculation method of the sign parameter is unknown, but it seems that it can be reused to some extent. See the following link.

- https://github.com/donavanbecker/homebridge-meross/issues/20#issuecomment-615852271

Therefore, by saving and reusing the credentials obtained from the MQTT announcement, you will be able to configure and use your device.